requirements.txt
flask
requests
beautifulsoup4
gunicorn
playwright
img2pdf
pillow
